% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/model.R
\name{Classifier}
\alias{Classifier}
\title{Linear Classifier specification constructor}
\usage{
Classifier(weights_init = NULL)
}
\arguments{
\item{weights_init}{initial weights.}
}
\value{
model.spec object
}
\description{
Linear Classifier specification constructor
}
\examples{
# generate initial weight matrix for a linear classifier with 10 classes
# and 784 features
w_init <- matrix(, 784, 10)
# Construct model
linear_classifier <- Classifier(weights_init=w_init)
}
\author{
Mohamed Ishmael Diwan Belghazi
}

