% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cudaLogReg.R
\docType{package}
\name{cudaLogReg}
\alias{cudaLogReg}
\alias{cudaLogReg-package}
\title{cudaLogReg: a GPU linear classfier for R}
\description{
This package implements a logistic regression with a batch gradient trainer.
Three backend are offered: R, C (BLAS), and CUDA
}

