% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/model_logreg.R
\name{get_condprob_logreg}
\alias{get_condprob_logreg}
\title{Computes Logistic regression class conditional probabilities}
\usage{
get_condprob_logreg(feats, weights, normalize = TRUE, log_domain = FALSE,
  backend = "R")
}
\arguments{
\item{feats}{features. (N feats M)}

\item{weights}{model weights. (M feats K)}

\item{normalize}{Normalize probabilties}

\item{log_domain}{Compute log-probabilties}

\item{backend}{Computation backend ("R", "C", or "CUDA")}
}
\value{
Class conditional probabilities
}
\description{
Computes Logistic regression class conditional probabilities
}
\author{
Mohamed Ishmael Diwan Belghazi
}

