% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/model_logreg.R
\name{get_cost_logreg}
\alias{get_cost_logreg}
\title{Computes cost function for logistic regression}
\usage{
get_cost_logreg(feats, weights, targets, decay = 0, backend = "R")
}
\arguments{
\item{feats}{Features (N feats M)}

\item{weights}{Model weights (M feats K)}

\item{targets}{Targets (N feats K)}

\item{decay}{L2 regularization decay factor}

\item{backend}{Computation backend ("R", "C", or "CUDA")}
}
\value{
logistic regression cost (cross-entropy)
}
\description{
Computes cost function for logistic regression
}
\author{
Mohamed Ishmael Diwan
}

