% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/generics.R
\name{get_error}
\alias{get_error}
\alias{get_error.model.spec}
\title{Computes model misclassification rate}
\usage{
get_error(object, feats, targets, backend = "R")

\method{get_error}{model.spec}(object, feats, targets, backend = "R")
}
\arguments{
\item{object}{Linear classifier specification object.}

\item{feats}{Numeric matrix. Features to predict. Dimensionality
should be consistent with that of the model.}

\item{targets}{Numeric matrix. TRUE targets to compare prediction against.
Dimensionality
should be consistent with that of the model. Should always be one-hot encoded.}

\item{backend}{Computation back-end ('R', 'C', or 'CUDA')}
}
\value{
Numeric scalar. Misclassification rate. Percentage of wrong predictions.
}
\description{
Computes model misclassification rate
}
\section{Methods (by class)}{
\itemize{
\item \code{model.spec}: Computes model misclassification rate for linear
classifier model object
}}
\author{
Mohamed Ishmael Diwan Belghazi
}

