% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/model_logreg.R
\name{get_grad_logreg}
\alias{get_grad_logreg}
\title{Computes gradient for logistic regression model}
\usage{
get_grad_logreg(feats, weights, targets, decay = 0, backend = "R")
}
\arguments{
\item{feats}{Features. (N feats M)}

\item{weights}{Model weights (M feats K)}

\item{targets}{One-hot encoded targets (N feats K)}

\item{decay}{L2 regularization decay factor}

\item{backend}{Computation backend ("R", "C", or "CUDA")}
}
\value{
gradient (M feats K)
}
\description{
Computes gradient for logistic regression model
}
\author{
Mohamed Ishmael Diwan Belghazi
}

