% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/gpuClassifieR.R
\docType{package}
\name{gpuClassifieR}
\alias{gpuClassifieR}
\alias{gpuClassifieR-package}
\title{gpuClassifieR: a GPU linear classifier for R}
\description{
This packages implements a linear classifier with a batch gradient descent trainer.
Three backend are offered: R, C, and CUDA.
}

