% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/specs.R
\name{LogReg}
\alias{LogReg}
\title{Logistic regression model specification}
\usage{
LogReg(weights_init = NULL)
}
\arguments{
\item{weights_init}{initial weights.}
}
\value{
model.spec object
}
\description{
Logistic regression model specification
}
\author{
Mohamed Ishmael Diwan Belghazi
}

