% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/utils.R
\name{logsumexp_R}
\alias{logsumexp_R}
\title{Log sum exp in R}
\usage{
logsumexp_R(x, ...)
}
\arguments{
\item{x}{vector to log sum exp}

\item{...}{pass through parameters}
}
\value{
log sum exp
}
\description{
Log sum exp in R
}
\author{
Mohamed Ishmael Diwan Belghazi
}

