% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cudaLogReg.R
\docType{data}
\name{mini_mnist}
\alias{mini_mnist}
\title{Mini-mnist}
\format{list of list of images and labels matrices}
\usage{
data(mini_mnist)
}
\description{
A dataset containing train and test Mnist images and labels
}
\keyword{datasets}

