% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/generics.R
\name{train}
\alias{train}
\alias{train.model.spec}
\title{Gradient Descent trainer generic}
\usage{
train(object, feats, targets, decay = NULL, step_size = NULL,
  max_iter = NULL, verbose = FALSE, tol = 1e-06, backend = "R", ...)

\method{train}{model.spec}(object, feats, targets, decay = NULL,
  step_size = NULL, max_iter = NULL, verbose = FALSE, tol = 1e-06,
  backend = "R", ...)
}
\arguments{
\item{object}{model.spec object}

\item{feats}{features (N X M)}

\item{targets}{targets (N X K)}

\item{decay}{L2 regularisation coefficient}

\item{step_size}{gradient descent step size}

\item{max_iter}{Maximum iteration}

\item{verbose}{prints output}

\item{tol}{assume convergence when norm of gradient is smaller than tol}

\item{backend}{Computation backend ("R", "C", or "CUDA")}

\item{...}{any other passthrough parameters}
}
\value{
model.spec object
}
\description{
Gradient Descent trainer generic
}
\section{Methods (by class)}{
\itemize{
\item \code{model.spec}: Gradient descent trainer
}}
\author{
Mohamed Ishmael Diwan Belghazi
}

