% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cudaLogReg.R
\docType{package}
\name{cudaLogReg}
\alias{cudaLogReg}
\alias{cudaLogReg-package}
\title{cudaLogReg: a GPU linear classfier for R}
\description{
This packages implements a linear classifier with a batch gradient descent trainer.
Three backend are offered: R, C, and CUDA.
}

