% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/generics.R
\name{get_cost}
\alias{get_cost}
\alias{get_cost.model.spec}
\title{Model cost accessor generic}
\usage{
get_cost(object, feats, targets, decay = NULL, backend = "R", ...)

\method{get_cost}{model.spec}(object, feats, targets, decay = NULL,
  backend = "R", ...)
}
\arguments{
\item{object}{object of class model.spec}

\item{feats}{Matrix of features (N X M)}

\item{targets}{matrix of one hot encoded target (N X K)}

\item{decay}{L2 regularisation decay}

\item{backend}{Computation backend ("R", "C", or "CUDA")}

\item{...}{Pass through parameters}
}
\value{
cost (M X K)
}
\description{
Model cost accessor generic
}
\section{Methods (by class)}{
\itemize{
\item \code{model.spec}: Model cost accessor
}}
\author{
Mohamed Ishmael Diwan Belghazi
}

