% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/model.R
\name{get_error}
\alias{get_error}
\alias{get_error.model.spec}
\title{Get model misclassification rate generic}
\usage{
get_error(object, feats, targets, backend = "R")

\method{get_error}{model.spec}(object, feats, targets, backend = "R")
}
\arguments{
\item{object}{Object of class model.spec}

\item{feats}{features to predict}

\item{targets}{Real targets}

\item{backend}{Computation backend ("R", "C", or "CUDA")}
}
\value{
Misclassification rate
}
\description{
Get model misclassification rate generic
}
\section{Methods (by class)}{
\itemize{
\item \code{model.spec}: Get model miscclassification rate
}}
\author{
Mohamed Ishmael Diwan Belghazi
}

