% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/model.R
\name{get_prob}
\alias{get_prob}
\alias{get_prob.model.spec}
\title{Class conditional probabilities getter generic}
\usage{
get_prob(object, feats, normalize = TRUE, log_domain = FALSE,
  backend = "R", ...)

\method{get_prob}{model.spec}(object, feats, normalize = TRUE,
  log_domain = FALSE, backend = "R", ...)
}
\arguments{
\item{object}{Object of class model.spec}

\item{feats}{features (N X M)}

\item{normalize}{normalize probabilities}

\item{log_domain}{return log probabilities}

\item{backend}{Computation backend ("R", "C", or "CUDA")}

\item{...}{Pass through parameters}
}
\value{
Class conditional probabilities
}
\description{
Class conditional probabilities getter generic
}
\section{Methods (by class)}{
\itemize{
\item \code{model.spec}: Class conditional probabilities getter
}}
\author{
Mohamed Ishmael Diwan Belghazi
}

