% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cudaLogReg.R
\docType{data}
\name{mini_mnist}
\alias{mini_mnist}
\title{Subset of MNIST}
\format{2 Matrices of 1000 rows and 784 columns for train and test images
and 2 matrices of 1000 rows and 10 columns for the train and test targets
(in onehot encoding).}
\usage{
data(mini_mnist)
}
\description{
A dataset containing a subset of thr MNIST database. The MNIST databse is a
large collection of handwritten digits.
}
\keyword{datasets}

